# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateAffiliateCodeRequest:
    properties:
      brand_commission_bps:
        description: Commission the brand earns per conversion (in bps of the order
          total)
        maximum: 5000
        minimum: 0
        type: integer
      code:
        maxLength: 50
        minLength: 3
        type: string
      per_user_limit:
        minimum: 0
        type: integer
      type:
        description: Promo semantics
        enum:
        - percentage
        - fixed
        type: string
      usage_limit:
        minimum: 0
        type: integer
      value_minor:
        minimum: 1
        type: integer
    required:
    - brand_commission_bps
    - code
    - type
    - value_minor
    type: object
  requests.CreateBrandProfileRequest:
    properties:
      business_name:
        maxLength: 255
        minLength: 2
        type: string
      business_size:
        enum:
        - startup
        - small
        - medium
        - large
        - enterprise
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_name:
        maxLength: 255
        minLength: 2
        type: string
      contact_phone:
        maxLength: 30
        minLength: 7
        type: string
      description:
        maxLength: 2000
        type: string
      industry:
        maxLength: 100
        minLength: 2
        type: string
      instagram:
        maxLength: 100
        type: string
      location:
        maxLength: 255
        minLength: 2
        type: string
      logo_url:
        maxLength: 500
        type: string
      preferred_categories:
        items:
          type: string
        maxItems: 20
        type: array
      tags:
        items:
          type: string
        maxItems: 20
        type: array
      target_audience:
        maxLength: 2000
        type: string
      website:
        maxLength: 500
        type: string
    required:
    - business_name
    - contact_name
    - industry
    - location
    type: object
...
  requests.UpdateBrandProfileRequest:
    properties:
      business_name:
        maxLength: 255
        minLength: 2
        type: string
      business_size:
        enum:
        - startup
        - small
        - medium
        - large
        - enterprise
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_name:
        maxLength: 255
        minLength: 2
        type: string
      contact_phone:
        maxLength: 30
        minLength: 7
        type: string
      description:
        maxLength: 2000
        type: string
      industry:
        maxLength: 100
        minLength: 2
        type: string
      instagram:
        maxLength: 100
        type: string
      is_active:
        type: boolean
      location:
        maxLength: 255
        minLength: 2
        type: string
      logo_url:
        maxLength: 500
        type: string
      preferred_categories:
        items:
          type: string
        maxItems: 20
        type: array
      tags:
        items:
          type: string
        maxItems: 20
        type: array
      target_audience:
        maxLength: 2000
        type: string
      website:
        maxLength: 500
        type: string
    type: object
...
  responses.AffiliateCodeResponse:
    properties:
      clicks:
        type: integer
      code:
        type: string
      commission_minor:
        type: integer
      conversions:
        type: integer
      created_at:
        type: string
      id:
        type: string
      partnership_id:
        type: string
      promo_code_id:
        type: string
      revenue_minor:
        type: integer
      updated_at:
        type: string
    type: object
...
  responses.BrandProfileResponse:
    properties:
      active_partnerships:
        description: Aggregated
        type: integer
      business_name:
        type: string
      business_size:
        type: string
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      industry:
        type: string
      instagram:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      logo_url:
        type: string
      preferred_categories:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      target_audience:
        type: string
      total_partnerships:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
      website:
        type: string
    type: object
...
  responses.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/responses.ErrorDetail'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@1kulture.com
    name: API Support
    url: https://1kulture.com/support
  description: Enterprise Event Management System API with multi-tenant SaaS architecture
  license:
    name: Proprietary
    url: https://1kulture.com/license
  termsOfService: https://1kulture.com/terms
  title: 1Kulture API
  version: "1.0"
paths:
...
  /brands:
    get:
      consumes:
      - application/json
      description: 'Browse brands. Filters: industry, location.'
      parameters:
      - description: Industry
        in: query
        name: industry
        type: string
      - description: Location
        in: query
        name: location
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Brands
          schema:
            items:
              $ref: '#/definitions/responses.BrandProfileResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List brand profiles
      tags:
      - brands
  /brands/{id}:
    get:
      consumes:
      - application/json
      description: Organizers can view brand profiles when evaluating partnership
        requests.
      parameters:
      - description: Brand Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand profile
          schema:
            $ref: '#/definitions/responses.BrandProfileResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a brand profile by ID
      tags:
      - brands
...
  /brands/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Brand profile
          schema:
            $ref: '#/definitions/responses.BrandProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my brand profile
      tags:
      - brands
    post:
      consumes:
      - application/json
      description: Create the brand profile for the authenticated user. Assigns the
        "brand" role.
      parameters:
      - description: Brand profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBrandProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Brand profile created
          schema:
            $ref: '#/definitions/responses.BrandProfileResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create brand profile
      tags:
      - brands
    put:
      consumes:
      - application/json
      parameters:
      - description: Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateBrandProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            $ref: '#/definitions/responses.BrandProfileResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my brand profile
      tags:
      - brands
...
  /partnerships/{id}/affiliate-codes:
    get:
      consumes:
      - application/json
      parameters:
      - description: Partnership ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Codes
          schema:
            items:
              $ref: '#/definitions/responses.AffiliateCodeResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List affiliate codes for a partnership
      tags:
      - affiliate-codes
    post:
      consumes:
      - application/json
      description: Organizer/co-organizer only. Creates a promo code tied to a partnership.
      parameters:
      - description: Partnership ID
        in: path
        name: id
        required: true
        type: string
      - description: Code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateAffiliateCodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.AffiliateCodeResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Code exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an affiliate code
      tags:
      - affiliate-codes
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"